Added yearly verification to CreditController

+++ src/main/java/pl/bolewski/credit_management/dto/BalanceDTO.java
@@ -4,10 +4,12 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Builder
 @Getter
 @Setter
 public class BalanceDTO {
-    private Long okoBalance;
-    private Long creditBalance;
+    private BigDecimal okoBalance;
+    private BigDecimal creditBalance;
 }
