+++ src/main/java/pl/bolewski/credit_management/dto/MoneyDTO.java
@@ -5,12 +5,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class MoneyDTO {
-    private Long cash;
+    private BigDecimal cash;
     private String account;
     private String month;
 }
